+++ depeche_db/_subscription.py
@@ -12,6 +12,7 @@ from typing import (
 from . import tools as _tools
 from ._aggregated_stream import AggregatedStream
 from ._interfaces import (
+    AckOpProtocol,
     CallMiddleware,
     ErrorAction,
     LockProvider,
@@ -59,7 +60,81 @@ class LogAndIgnoreSubscriptionErrorHandler(SubscriptionErrorHandler):
         return ErrorAction.IGNORE
 
 
+class NoAckOp:
+    def execute(self, **kwargs):
+        raise RuntimeError(
+            "NoAckOp cannot be executed, choose a different Ack strategy to use this operation"
+        )
+
+
+class AckRolledback(Exception):
+    pass
+
+
+class AckOp:
+    def __init__(
+        self,
+        name: str,
+        partition: int,
+        position: int,
+        state_provider: SubscriptionStateProvider,
+    ):
+        self.name = name
+        self.partition = partition
+        self.position = position
+        self._state_provider = state_provider
+        self._executed = False
+        self._rolled_back = False
+
+    def execute(self, **subscription_state_provider_kwargs):
+        if self._executed:
+            self._check()
+            return
+
+        if subscription_state_provider_kwargs:
+            provider = self._state_provider.session(
+                **subscription_state_provider_kwargs
+            )
+        else:
+            provider = self._state_provider
+
+        self._executed = True
+        state = provider.read(self.name)
+        assert (
+            state.positions.get(self.partition, -1) == self.position - 1
+        ), f"{self.partition} should have {self.position - 1} as last position, but has {state.positions.get(self.partition, -1)}"
+        provider.store(
+            subscription_name=self.name,
+            partition=self.partition,
+            position=self.position,
+        )
+
+    def _check(self):
+        state = self._state_provider.read(self.name)
+        if state.positions.get(self.partition, -1) != self.position:
+            raise AckRolledback()
+
+    def rollback(self):
+        self._rolled_back = True
+
+    @property
+    def rolled_back(self):
+        return self._rolled_back
+
+
+class AckStrategy:
+    def get_op(self) -> AckOpProtocol:
+        raise NotImplementedError()
+
+
+class ManagedAckStrategy(AckStrategy):
+    def get_op(self) -> AckOpProtocol:
+        return NoAckOp()
+
+
 class Subscription(Generic[E]):
+    _state_provider: SubscriptionStateProvider
+
     def __init__(
         self,
         name: str,
@@ -159,27 +234,30 @@ class Subscription(Generic[E]):
                     }
 
                     for pointer in message_pointers:
+                        ack = AckOp(
+                            name=self.name,
+                            partition=pointer.partition,
+                            position=pointer.position,
+                            state_provider=self._state_provider,
+                        )
+
                         yield SubscriptionMessage(
                             partition=pointer.partition,
                             position=pointer.position,
                             stored_message=stored_messages[pointer.message_id],
+                            ack=ack,
                         )
-                        self._ack(
-                            partition=pointer.partition,
-                            position=pointer.position,
-                        )
+                        if ack.rolled_back:
+                            # the message was not ack'd or the acknolwedgement was rolled back
+                            break
+                        try:
+                            ack.execute()
+                        except AckRolledback:
+                            # the message was not ack'd or the acknolwedgement was rolled back
+                            break
                 break
             finally:
                 self._lock_provider.unlock(lock_key)
-
-    def _ack(self, partition: int, position: int):
-        state = self._state_provider.read(self.name)
-        assert (
-            state.positions.get(partition, -1) == position - 1
-        ), f"{partition} should have {position - 1} as last position, but has {state.positions.get(partition, -1)}"
-        self._state_provider.store(
-            subscription_name=self.name, partition=partition, position=position
-        )
 
 
 class SubscriptionMessageHandler(Generic[E]):