+++ tests/_tools.py
@@ -1,4 +1,5 @@
 import uuid as _uuid
+from urllib import parse as _urlparse
 
 import sqlalchemy as _sa
 
@@ -6,23 +7,18 @@ import sqlalchemy as _sa
 def pg_check_if_db_exists(uri: str) -> bool:
     with pg_get_connection(uri, use_postgres_db=True) as conn:
         db_name = _get_db_name(uri)
-        return (  # type: ignore
-            conn.execute(
-                _sa.text("select count(*) from pg_database where datname = :db_name"),
-                {"db_name": db_name},
-            ).scalar()
-            > 0
-        )
+        count: int = conn.execute(
+            _sa.text("select count(*) from pg_database where datname = :db_name"),
+            {"db_name": db_name},
+        ).scalar()
+        return count > 0
 
 
 def pg_create_db(uri: str):
     db_name = _get_db_name(uri)
     print(f"Creating database {db_name}")
     with pg_get_connection(uri, use_postgres_db=True) as conn:
-        conn.execute(_sa.text("ROLLBACK"))
-        # transaction = conn.begin()
         conn.execute(_sa.text(f"CREATE DATABASE {db_name}"))
-        # transaction.commit()
         print(f"Created database {db_name}")
 
 
@@ -30,22 +26,17 @@ def pg_drop_db(uri: str):
     db_name = _get_db_name(uri)
     print(f"Dropping database {db_name}")
     with pg_get_connection(uri, use_postgres_db=True) as conn:
-        conn.execute(_sa.text("ROLLBACK"))
-        # transaction = conn.begin()
         conn.execute(_sa.text(f"DROP DATABASE {db_name}"))
-        # transaction.commit()
         print(f"Dropped database {db_name}")
 
 
-def pg_get_connection(uri: str, use_postgres_db: bool = False) -> _sa.engine.Connection:
-    from urllib.parse import urlparse, urlunparse
-
+def pg_get_connection(uri: str, use_postgres_db: bool = False):
+    parts = list(_urlparse.urlparse(uri))
     if use_postgres_db:
-        parts = list(urlparse(uri))
         parts[2] = "/postgres"
-        uri = urlunparse(parts)
+    uri = _urlparse.urlunparse(parts)
 
-    return _sa.create_engine(uri).connect()
+    return _sa.create_engine(uri, isolation_level="AUTOCOMMIT").connect()
 
 
 def _get_db_name(uri: str) -> str: